+++ nata-monkey/src/main/java/com/nata/AllRunner.java
@@ -1,6 +1,7 @@
 package com.nata;
 
 import com.nata.cmd.AdbDevice;
+import com.nata.monkeys.QLearningMonkey;
 import com.nata.monkeys.RandomMonkey;
 
 /**
@@ -8,17 +9,16 @@ import com.nata.monkeys.RandomMonkey;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-19 16:15
  */
-public class RMRunner {
+public class AllRunner {
     public static void main(String[] args) {
-        //Info about the apk to test
-//        String pkg = "com.zhihu.android";
-//        String act = ".app.ui.activity.MainActivity";
 
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
+
         //Set Monkey settings
         AdbDevice device = new AdbDevice();
         RandomMonkey randomMonkey = new RandomMonkey(pkg,act,device);
+        QLearningMonkey qLearningMonkey= new QLearningMonkey(pkg,act,device);
 
         for(int i = 0; i  < 3; i++){
             //Start run monkey
@@ -26,6 +26,14 @@ public class RMRunner {
 
             //Do report
             randomMonkey.report();
+        }
+
+        for(int i = 0 ; i < 3 ; i++){
+            //Start run monkey
+            qLearningMonkey.play();
+
+            //Do report
+            qLearningMonkey.report();
         }
 
     }