+++ nata-monkey/src/main/java/com/nata/action/ActionType.java
@@ -14,6 +14,7 @@ public class ActionType {
     public static final String INPUT= "TextInputAction";
     public static final String BACK= "BackAction";
     public static final String MENU= "MenuAction";
+    public static final String HOME= "HomeAction";
     public static final String START_APP= "StartAppAction";