+++ nata-monkey/src/main/java/com/nata/monkeys/RandomMonkey.java
@@ -3,10 +3,9 @@ package com.nata.monkeys;
 import com.nata.action.*;
 import com.nata.cmd.AdbDevice;
 import com.nata.element.Widget;
+import com.nata.state.State;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
 /**
  * Author: Calvin Meng
@@ -17,140 +16,126 @@ public class RandomMonkey extends AbstractMonkey {
     volatile private boolean isRunning = true;
     private List<Action> actionList = new ArrayList<>();
     private final int ACTION_LIMIT = 20;
+    private final int ACTION_COUNTS = 200;
     private final double P_ENTITY = 0.2;
     private final double P_SWIPE = 0.2;
     private final Random random  = new Random();
     private Action restartAction = null;
     private Action backAction = null;
     private Action menuAction = null;
+    private Action homeAction = null;
+
     private Action lastAction = null;
     private ActionFactory actionFactory = null;
+    private Set<String> activitySet = new HashSet<>();
 
     public RandomMonkey(String pkg, String act, AdbDevice device) {
         super("randomMonkey", pkg, act, device);
         actionFactory = new ActionFactory(device);
         restartAction = actionFactory.CreateRestartAction(getPkgAct());
         backAction = actionFactory.createBackAction();
+        homeAction = actionFactory.createHomeAction();
     }
 
+    private void startApp() {
+        System.out.println("Random Monkey start playing...");
+        clearAppData();
+        System.out.println("cleaning app data...");
 
-    private class UINodeAction {
-        Widget node;
-        String Action;
+        restartAction.fire();
+        actionList.add(restartAction);
+        System.out.println("starting app success...");
+    }
 
-        UINodeAction(Widget node, String Action) {
-            this.node = node;
-            this.Action = Action;
+    private State getBackToApp() {
+        boolean forceQuit = false;
+        while (!isInCurrentPkg()) {
+            // if even the restart action cannot restart it ;
+            if(forceQuit){
+                clearAppData();
+                restartAction.fire();
+                actionList.add(restartAction);
+                lastAction = restartAction;
+            }
+            else if(lastAction instanceof StartAppAction){
+                homeAction.fire();
+                actionList.add(homeAction);
+                restartAction.fire();
+                actionList.add(restartAction);
+                lastAction = restartAction;
+                forceQuit = true;
+            }
+            // if monkey get out of pkg because of back action
+            else if (lastAction instanceof BackAction) {
+                restartAction.fire();
+                actionList.add(restartAction);
+                lastAction = restartAction;
+            } else {
+                backAction.fire();
+                actionList.add(backAction);
+                if (!isInCurrentPkg()) {
+                    backAction.fire();
+                    actionList.add(backAction);
+                }
+                lastAction = backAction;
+            }
+        }
+        // get current state
+        State state = getCurrentState();
+        return state;
+    }
+
+    private Action chooseActionFromState(State curState) {
+        Map<Action,Double> actionTable = actionFactory.getActionsFromState(curState);
+        if (actionTable == null || actionTable.size() == 0) {
+            System.out.println("Error cannot get actionTable or no action can be choosed");
+            return backAction;
         }
 
-        public Widget getNode() {
-            return node;
-        }
-
-        public String getAction() {
-            return Action;
-        }
+        Set<Action> actionSet = actionTable.keySet();
+        List<Action> list = new ArrayList<>();
+        list.addAll(actionSet);
+        Collections.shuffle(list);
+        return list.get(0);
     }
 
     @Override
     public void play() {
-        System.out.println("Random Monkey start playing...");
-        restartAction.fire();
-        actionList.add(restartAction);
+//        System.out.println("Random Monkey start playing...");
+//        restartAction.fire();
+//        actionList.add(restartAction);
 
-        while (isRunning) {
-            Action nextAction;
-            // if not in current pkg
-            if(!isInCurrentPkg()){
-                // if monkey get out of pkg because of back action
-                if(lastAction instanceof BackAction){
-                    nextAction = restartAction;
-                }else {
-                    nextAction = backAction;
-                }
-            }else{
-                nextAction= getNextAction();
-            }
+        startApp();
 
-            if(nextAction == null){
+        State curState = getCurrentState();
+        if (isInCurrentPkg()) {
+            activitySet.add(curState.getActivity());
+        }
+
+        int cnt = 0;
+        while ((++cnt) <= ACTION_COUNTS) {
+            // if not in pkg, get it back
+            if (!isInCurrentPkg()) {
+                curState = getBackToApp();
                 continue;
             }
 
+            Action action = chooseActionFromState(curState);
+            actionList.add(action);
 
-            actionList.add(nextAction);
-            nextAction.fire();
-            lastAction = nextAction;
+            action.fire();
+            lastAction = action;
 
-            System.out.println(nextAction);
-            if(isCrashed()){
-                System.out.println("The App Crashed");
-                break;
-            }
-
-            if (actionList.size() >= ACTION_LIMIT) {
-                actionList.clear();
-                restartAction.fire();
-                actionList.add(restartAction);
+            curState = getCurrentState();
+            if (isInCurrentPkg()) {
+                activitySet.add(curState.getActivity());
             }
         }
-        report();
+//        report();
     }
 
 
-    public Action getNextAction() {
-        //First Gamble : Whether to Take Functional Actions: Back & Menu
-        double randomValue = Math.random();
-        if(randomValue < P_ENTITY){
-            double gamble = Math.random();
-            if(gamble > 0.5){
-                return backAction;
-            }else{
-                return menuAction;
-            }
-            //Second Gamble : Take Swipe Actions
-        }
-//        else if ( (randomValue - P_ENTITY) <= P_SWIPE){
-//            int gamble  = random.nextInt(4);
-//            switch (gamble){
-//                case 0: actionFactory.CreateSwipeAction(SwipeDirection.LEFT);break;
-//                case 1: actionFactory.CreateSwipeAction(SwipeDirection.RIGHT);break;
-//                case 2: actionFactory.CreateSwipeAction(SwipeDirection.UP);break;
-//                case 3: actionFactory.CreateSwipeAction(SwipeDirection.DOWN);break;
-//            }
-//        }
 
-        //Action nextAction;
-
-        List<Widget> list = GrabCurrentUi();
-        List<UINodeAction> actionList = new ArrayList<>();
-        if (list != null) {
-            for (Widget node : list) {
-                if (node.getClassName().equals("android.widget.Button") && node.getClickable().equals("true")) {
-                    actionList.add(new UINodeAction(node, ActionType.TAP));
-                }
-                if (node.getClassName().equals("android.widget.EditText") && node.getText().length() == 0) {
-                    actionList.add(new UINodeAction(node, ActionType.INPUT));
-                }
-            }
-            if (actionList.size() > 0) {
-                int randomIndex = (int) (Math.random() * actionList.size());
-                UINodeAction nodeAction = actionList.get(randomIndex);
-                Action action = null;
-                switch (nodeAction.getAction()) {
-                    case ActionType.INPUT:
-                        action = new TextInputAction(getDevice(),nodeAction.getNode());
-                        break;
-                    case ActionType.TAP:
-                        action = new TapAction( getDevice(),nodeAction.getNode());
-                        break;
-                }
-                return action;
-            } else {
-                return null;
-            }
-        }
-        return null;
-    }
 
     @Override
     public void stop() {
@@ -159,8 +144,24 @@ public class RandomMonkey extends AbstractMonkey {
 
     @Override
     public void report() {
-        for (Action action:actionList) {
-           System.out.println(action);
+        System.out.println("--------------------[Activities report]--------------------");
+        for (String activity : activitySet) {
+            System.out.println(activity);
+        }
+
+        System.out.println("--------------------[Widget report]--------------------");
+        Set<Widget> widgetSet = getWidgetSet();
+        int widgetCount = widgetSet.size();
+        System.out.println("Widget count: " + widgetCount);
+        for (Widget node : widgetSet) {
+            System.out.println(node);
+        }
+
+        System.out.println("--------------------[Action report]--------------------");
+        int actionCount = actionList.size();
+        System.out.println("Actions count: " + actionCount);
+        for (Action action : actionList) {
+            System.out.println(action);
         }
     }